requirements.txt
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.5.0
